[refactor] move classes in model folder to reflect on UML class diagram
