+++ src/main/java/scs/comp5903/cucumber/builder/JStepParameterExtractor.java
@@ -1,11 +1,11 @@
 package scs.comp5903.cucumber.builder;
 
 import org.slf4j.Logger;
-import scs.comp5903.cucumber.model.JStepDefMethodDetail;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
 import scs.comp5903.cucumber.model.exception.ErrorCode;
-import scs.comp5903.cucumber.model.jstep.*;
-import scs.comp5903.cucumber.model.matcher.*;
+import scs.comp5903.cucumber.model.feature.jstep.*;
+import scs.comp5903.cucumber.model.stepdef.JStepDefMethodDetail;
+import scs.comp5903.cucumber.model.stepdef.matcher.*;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
