+++ src/main/java/scs/comp5903/cucumber/model/feature/JScenarioOutlineDetail.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.model;
+package scs.comp5903.cucumber.model.feature;
 
 import java.util.List;
 import java.util.Objects;
