+++ src/main/java/scs/comp5903/cucumber/model/stepdef/JStepDefDetail.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.model;
+package scs.comp5903.cucumber.model.stepdef;
 
 import java.util.List;
 import java.util.Objects;
