+++ src/main/java/scs/comp5903/cucumber/model/stepdef/matcher/AbstractJStepMatcher.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.model.matcher;
+package scs.comp5903.cucumber.model.stepdef.matcher;
 
 
 import java.util.Objects;
