+++ src/main/java/scs/comp5903/cucumber/model/stepdef/matcher/ButJStepMatcher.java
@@ -1,11 +1,11 @@
-package scs.comp5903.cucumber.model.matcher;
+package scs.comp5903.cucumber.model.stepdef.matcher;
 
 /**
  * @author Charles Chen 101035684
  * @date 2022-06-22
  */
-public class AndJStepMatcher extends AbstractJStepMatcher {
-  public AndJStepMatcher(String matchingString) {
+public class ButJStepMatcher extends AbstractJStepMatcher {
+  public ButJStepMatcher(String matchingString) {
     super(matchingString);
   }
 
@@ -14,10 +14,10 @@ public class AndJStepMatcher extends AbstractJStepMatcher {
     if (o == this) {
       return true;
     }
-    if (!(o instanceof AndJStepMatcher)) {
+    if (!(o instanceof ButJStepMatcher)) {
       return false;
     }
-    final AndJStepMatcher other = (AndJStepMatcher) o;
+    final ButJStepMatcher other = (ButJStepMatcher) o;
     if (!other.canEqual(this)) {
       return false;
     }
@@ -26,7 +26,7 @@ public class AndJStepMatcher extends AbstractJStepMatcher {
 
   @Override
   protected boolean canEqual(final Object other) {
-    return other instanceof AndJStepMatcher;
+    return other instanceof ButJStepMatcher;
   }
 
   @Override
@@ -36,6 +36,6 @@ public class AndJStepMatcher extends AbstractJStepMatcher {
 
   @Override
   public String toString() {
-    return "AndJStepMatcher(matchingString=" + this.getMatchingString() + ")";
+    return "ButJStepMatcher(matchingString=" + this.getMatchingString() + ")";
   }
 }
