+++ src/main/java/scs/comp5903/cucumber/parser/DetailBuilder.java
@@ -1,11 +1,11 @@
 package scs.comp5903.cucumber.parser;
 
 import org.slf4j.Logger;
-import scs.comp5903.cucumber.model.JScenarioDetail;
-import scs.comp5903.cucumber.model.JScenarioOutlineDetail;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
 import scs.comp5903.cucumber.model.exception.ErrorCode;
-import scs.comp5903.cucumber.model.jstep.*;
+import scs.comp5903.cucumber.model.feature.JScenarioDetail;
+import scs.comp5903.cucumber.model.feature.JScenarioOutlineDetail;
+import scs.comp5903.cucumber.model.feature.jstep.*;
 
 import java.util.ArrayList;
 import java.util.Arrays;
