+++ src/main/java/scs/comp5903/cucumber/parser/JFeatureFileLineByLineParser.java
@@ -1,9 +1,9 @@
 package scs.comp5903.cucumber.parser;
 
 import org.slf4j.Logger;
-import scs.comp5903.cucumber.model.JFeatureDetail;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
 import scs.comp5903.cucumber.model.exception.ErrorCode;
+import scs.comp5903.cucumber.model.feature.JFeatureDetail;
 import scs.comp5903.cucumber.util.JFeatureKeyword;
 import scs.comp5903.cucumber.util.LineUtil;
 import scs.comp5903.cucumber.util.ThrowingConsumer;
