+++ src/test/java/scs/comp5903/cucumber/builder/JStepParameterExtractorTest.java
@@ -1,10 +1,10 @@
 package scs.comp5903.cucumber.builder;
 
 import org.junit.jupiter.api.Test;
-import scs.comp5903.cucumber.model.JStepDefMethodDetail;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
-import scs.comp5903.cucumber.model.jstep.GivenStep;
-import scs.comp5903.cucumber.model.matcher.GivenJStepMatcher;
+import scs.comp5903.cucumber.model.feature.jstep.GivenStep;
+import scs.comp5903.cucumber.model.stepdef.JStepDefMethodDetail;
+import scs.comp5903.cucumber.model.stepdef.matcher.GivenJStepMatcher;
 
 import static org.junit.jupiter.api.Assertions.*;
