+++ src/test/java/scs/comp5903/cucumber/parser/DetailBuilderTest.java
@@ -2,7 +2,7 @@ package scs.comp5903.cucumber.parser;
 
 import org.junit.jupiter.api.Test;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
-import scs.comp5903.cucumber.model.jstep.*;
+import scs.comp5903.cucumber.model.feature.jstep.*;
 
 import java.util.List;
