+++ src/test/java/scs/comp5903/cucumber/parser/LineByLineParserParsingTest.java
@@ -2,8 +2,8 @@ package scs.comp5903.cucumber.parser;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import scs.comp5903.cucumber.model.JFeatureDetail;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
+import scs.comp5903.cucumber.model.feature.JFeatureDetail;
 
 import java.util.ArrayList;
 import java.util.List;
