+++ src/test/java/scs/comp5903/cucumber/sample/CmsSimpleRegisterCourseStepDefs.java
@@ -11,7 +11,7 @@ import scs.comp5903.cucumber.sample.cmsutil.SeleniumFactory;
 import java.util.concurrent.CountDownLatch;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static scs.comp5903.cucumber.model.JStepKeyword.*;
+import static scs.comp5903.cucumber.model.annotation.step.JStepKeyword.*;
 
 /**
  * @author Charles Chen 101035684
