+++ src/test/java/scs/comp5903/cucumber/sample/DecimalAndBigIntStepDef.java
@@ -1,8 +1,8 @@
 package scs.comp5903.cucumber.sample;
 
 import org.slf4j.Logger;
-import scs.comp5903.cucumber.model.JStepKeyword;
 import scs.comp5903.cucumber.model.annotation.step.JStep;
+import scs.comp5903.cucumber.model.annotation.step.JStepKeyword;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
