+++ src/test/java/scs/comp5903/cucumber/sample/RummikubDummyJStepDefs.java
@@ -3,7 +3,7 @@ package scs.comp5903.cucumber.sample;
 import org.slf4j.Logger;
 import scs.comp5903.cucumber.model.annotation.step.JStep;
 
-import static scs.comp5903.cucumber.model.JStepKeyword.*;
+import static scs.comp5903.cucumber.model.annotation.step.JStepKeyword.*;
 
 public class RummikubDummyJStepDefs {
   private final static Logger log = org.slf4j.LoggerFactory.getLogger(RummikubDummyJStepDefs.class);
