+++ src/test/java/scs/comp5903/cucumber/sample/samplestepdef/RummikubDummySeparatedJStepDefs.java
@@ -3,7 +3,7 @@ package scs.comp5903.cucumber.sample.samplestepdef;
 import org.slf4j.Logger;
 import scs.comp5903.cucumber.model.annotation.step.JStep;
 
-import static scs.comp5903.cucumber.model.JStepKeyword.*;
+import static scs.comp5903.cucumber.model.annotation.step.JStepKeyword.*;
 
 /**
  * just exactly same step definition of {@link scs.comp5903.cucumber.sample.RummikubDummyJStepDefs}, but separated into different class
